+++ CHANGELOG.md
@@ -116,3 +116,4 @@ was removed or the new one was added to folder
 - Textures now store they last write date
 - Reloading textures in Texture Manager now also puts them to they newset state on PC
 - Colors can be initialized with enum
+- CreateTempFont from TTF repaired and should be used as de facto standard instead of create font